change all integer sql metadata to long

+++ src/main/kotlin/com/young_zy/forum/controller/request/PostThreadRequest.kt
@@ -1,7 +1,7 @@
 package com.young_zy.forum.controller.request
 
 data class PostThreadRequest(
-        val sectionId: Int,
+        val sectionId: Long,
         val title: String,
         val content: String,
         val isQuestion: Boolean = false
